requirements.txt
pywin32
PyOpenGL
imgui[full]
glfw